RAG_CHATBOT/RAG_CHATBOT/requirements.txt
flask==2.3.3
openai==0.28.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
cohere==4.33
boto3==1.28.22
supabase==2.0.0
pyngrok==7.0.0
